+++ lib/ArduinoJson/ArduinoJson.h
@@ -5,11 +5,12 @@
 // package of ArduinoJson.
 // In that case, just delete the current installation and install the package.
 
-// Copyright Benoit Blanchon 2014
+// Copyright Benoit Blanchon 2014-2016
 // MIT License
 //
 // Arduino JSON library
 // https://github.com/bblanchon/ArduinoJson
+// If you like this project, please add a star!
 
 #include "include/ArduinoJson/DynamicJsonBuffer.hpp"
 #include "include/ArduinoJson/JsonArray.hpp"